pyproject.toml
[project]
name = "parsehub"
version = "1.5.14"
description = "支持AI总结的社交媒体聚合解析器"
readme = "README.md"
requires-python = ">=3.12.0"
authors = [
    { name = "梓澪", email = "zilingmio@gmail.com" }
]
license = { text = "MIT" }
dependencies = [
    "aiofiles>=23.2",
    "beautifulsoup4>=4.12.3",
    "loguru>=0.6.0",
    "opencv-python>=4.10.0.84",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "tenacity>=8.5.0",
    "urlextract>=1.9.0",
    "yt-dlp[default]",
    "lxml>=5.3.0",
    "instaloader>=4.14",
    "pydantic>=1.10.19",
    "markdownify>=1.1.0",
    "markdown>=3.7",
    "requests",
    "httpx>=0.24.1",
    "pydantic-settings>=2.10.1",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.hatch.build.targets.wheel]
packages = ["src/parsehub"]

[project.urls]
Repository = "https://github.com/z-mio/parsehub"
Issues = "https://github.com/z-mio/parsehub/issues"

[tool.ruff]
line-length = 120
exclude = [
    "test",
    "src/parsehub/deps/xhs"
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle 错误检查
    "W", # pycodestyle 警告检查
    "F", # pyflakes 错误检查
    "I", # isort 导入排序
    "B", # flake8-bugbear 常见错误检查
    "C4", # flake8-comprehensions 列表/字典推导式检查
    "UP", # pyupgrade 自动升级语法
]
ignore = [
    "B008", # 不在参数默认值中执行函数调用
    "C901", # 函数复杂度过高
]

